refactor: extract macro types and parsing into separate crates
Move unsynn grammar definitions from facet-macros-impl into facet-macro-types,
and parsed type representations (PStruct, PEnum, etc.) into facet-macro-parse.

This enables the plugin system POC to depend on just facet-macro-parse for
type parsing without pulling in the entire codegen machinery.

- facet-macro-types: unsynn grammar, keywords, operators, RenameRule
- facet-macro-parse: PStruct, PEnum, PContainer, PAttrs, generic_params, unescaping
- facet-macros-impl: now re-exports from both crates
