+++ facet-macro-types/Cargo.toml
@@ -5,10 +5,11 @@ edition.workspace = true
 rust-version.workspace = true
 license.workspace = true
 repository.workspace = true
-description = "Shared parsed type representations for facet macro ecosystem"
+description = "Unsynn grammar and type definitions for facet macro ecosystem"
 keywords = ["reflection", "macros", "derive"]
 categories = ["development-tools"]
 
 [dependencies]
+unsynn = "0.3.0"
 proc-macro2 = "1"
 quote = "1"
