+++ facet-macros-impl/Cargo.toml
@@ -25,6 +25,8 @@ doc = []
 helpful-derive = ["dep:strsim"]
 
 [dependencies]
+facet-macro-types = { version = "0.1.0", path = "../facet-macro-types" }
+facet-macro-parse = { version = "0.1.0", path = "../facet-macro-parse" }
 unsynn = "0.3.0"
 proc-macro2 = "1.0.95"
 quote = "1.0.40"
