+++ poc/plugin-poc-test/src/main.rs
@@ -22,7 +22,9 @@ pub enum MyError {
 #[derive(FacetPoc)]
 #[facet_poc(display, debug)]
 pub struct Point {
+    #[allow(dead_code)]
     x: i32,
+    #[allow(dead_code)]
     y: i32,
 }
 
@@ -43,4 +45,49 @@ fn main() {
     println!("Point debug: {p:?}");
 
     println!("\nPlugin system POC works!");
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_enum_display_unit_variant() {
+        let err = MyError::Unknown;
+        assert_eq!(format!("{err}"), "unknown error occurred");
+    }
+
+    #[test]
+    fn test_enum_display_struct_variant() {
+        let err = MyError::InvalidHeader {
+            expected: "JSON".to_string(),
+            found: "XML".to_string(),
+        };
+        // The field names in the doc comment get interpolated!
+        assert_eq!(
+            format!("{err}"),
+            "invalid header (expected JSON, found XML)"
+        );
+    }
+
+    #[test]
+    fn test_enum_display_tuple_variant() {
+        let err = MyError::Disconnect(std::io::Error::new(
+            std::io::ErrorKind::ConnectionRefused,
+            "refused",
+        ));
+        assert_eq!(format!("{err}"), "data store disconnected");
+    }
+
+    #[test]
+    fn test_struct_display() {
+        let p = Point { x: 10, y: 20 };
+        assert_eq!(format!("{p}"), "A simple point");
+    }
+
+    #[test]
+    fn test_struct_debug() {
+        let p = Point { x: 10, y: 20 };
+        assert_eq!(format!("{p:?}"), "Point");
+    }
 }
