+++ poc/plugin-poc/src/lib.rs
@@ -9,31 +9,25 @@
 //! 3. Each plugin adds its "template" and forwards to the next
 //! 4. `__facet_finalize!` parses once and generates all code
 //!
-//! For this POC, we simulate the templating with direct codegen.
-//! Uses unsynn for parsing instead of syn.
-
-mod template;
+//! Now using the proper facet-macro-* crates for parsing and templating.
 
 use proc_macro::TokenStream;
 use proc_macro2::TokenStream as TokenStream2;
 use quote::quote;
 
-// Import unsynn but avoid shadowing std::result::Result
-use unsynn::{
-    At, BraceGroupContaining, BracketGroupContaining, Colon, Comma, Gt, Ident, Literal, Lt,
-    ParenthesisGroupContaining, Pound, TokenTree, keyword, unsynn,
-};
+// Use the new crates
+use facet_macro_parse::{PType, parse_type};
+
+// Import unsynn for the internal protocol parsing
+use unsynn::{At, BraceGroupContaining, Comma, Ident, Literal, keyword, unsynn};
 use unsynn::{IParse, ToTokenIter};
 
 // Define keywords we need
 keyword! {
     KFacetPoc = "facet_poc";
-    KPub = "pub";
-    KEnum = "enum";
-    KStruct = "struct";
 }
 
-// Grammar for parsing the internal macro invocation format (private, not pub)
+// Grammar for parsing the internal macro invocation format
 unsynn! {
     /// Section marker like `@tokens`, `@remaining`, `@plugins`
     struct SectionMarker {
@@ -45,6 +39,12 @@ unsynn! {
     struct BracedSection {
         marker: SectionMarker,
         content: BraceGroupContaining<TokenStream2>,
+    }
+
+    /// Attribute like `#[...]`
+    struct Attribute {
+        pound: unsynn::Pound,
+        body: unsynn::BracketGroupContaining<TokenStream2>,
     }
 }
 
@@ -69,10 +69,8 @@ impl PluginInvoke {
                     tokens = Some(section.content.content);
                 }
                 "remaining" => {
-                    // Parse comma-separated paths - just collect token trees
                     let content = section.content.content;
                     if !content.is_empty() {
-                        // Split by comma and collect each path
                         let mut current = TokenStream2::new();
                         for tt in content {
                             if let proc_macro2::TokenTree::Punct(p) = &tt {
@@ -92,16 +90,12 @@ impl PluginInvoke {
                     }
                 }
                 "plugins" => {
-                    // Parse comma-separated string literals
                     let content = section.content.content;
                     let mut inner = content.to_token_iter();
                     while let Ok(lit) = inner.parse::<Literal>() {
-                        // Extract string value from literal
                         let s = lit.to_string();
-                        // Remove quotes
                         let unquoted = s.trim_matches('"');
                         plugins.push(unquoted.to_string());
-                        // Skip comma if present
                         let _ = inner.parse::<Comma>();
                     }
                 }
@@ -119,190 +113,6 @@ impl PluginInvoke {
     }
 }
 
-/// Parsed struct/enum for code generation
-struct ParsedType {
-    name: Ident,
-    kind: TypeKind,
-}
-
-enum TypeKind {
-    Enum(Vec<ParsedVariant>),
-    Struct { doc: String },
-}
-
-struct ParsedVariant {
-    name: Ident,
-    doc: String,
-    fields: VariantFields,
-}
-
-enum VariantFields {
-    Unit,
-    Tuple(usize),
-    Named(Vec<Ident>),
-}
-
-// Grammar for attributes and visibility (private)
-unsynn! {
-    /// Attribute like `#[doc = "..."]` or `#[facet_poc(display)]`
-    struct Attribute {
-        pound: Pound,
-        body: BracketGroupContaining<TokenStream2>,
-    }
-
-    /// Visibility
-    enum Vis {
-        Pub(KPub),
-    }
-
-    /// Generics like `<T, U>`
-    struct Generics {
-        lt: Lt,
-        content: TokenStream2,
-        gt: Gt,
-    }
-}
-
-impl ParsedType {
-    fn parse(tokens: TokenStream2) -> Result<Self, String> {
-        let mut iter = tokens.clone().to_token_iter();
-
-        // Skip attributes and visibility
-        while iter.parse::<Attribute>().is_ok() {}
-        let _ = iter.parse::<Vis>();
-
-        // Check if enum or struct
-        if iter.parse::<KEnum>().is_ok() {
-            let name: Ident = iter
-                .parse()
-                .map_err(|e| format!("expected enum name: {e}"))?;
-            // Skip generics if present
-            let _ = iter.parse::<Generics>();
-
-            let body: BraceGroupContaining<TokenStream2> = iter
-                .parse()
-                .map_err(|e| format!("expected enum body: {e}"))?;
-
-            let variants = Self::parse_enum_variants(body.content)?;
-
-            Ok(ParsedType {
-                name,
-                kind: TypeKind::Enum(variants),
-            })
-        } else if iter.parse::<KStruct>().is_ok() {
-            // For struct, just get the name and doc
-            let name: Ident = iter
-                .parse()
-                .map_err(|e| format!("expected struct name: {e}"))?;
-
-            // Re-parse to get doc from attributes
-            let mut doc_iter = tokens.to_token_iter();
-            let mut doc = String::new();
-            while let Ok(attr) = doc_iter.parse::<Attribute>() {
-                if let Some(d) = extract_doc(&attr) {
-                    if !doc.is_empty() {
-                        doc.push(' ');
-                    }
-                    doc.push_str(&d);
-                }
-            }
-
-            Ok(ParsedType {
-                name,
-                kind: TypeKind::Struct { doc },
-            })
-        } else {
-            Err("expected enum or struct".to_string())
-        }
-    }
-
-    fn parse_enum_variants(body: TokenStream2) -> Result<Vec<ParsedVariant>, String> {
-        let mut variants = Vec::new();
-        let mut iter = body.to_token_iter();
-
-        loop {
-            // Collect doc comments for this variant
-            let mut doc = String::new();
-            while let Ok(attr) = iter.parse::<Attribute>() {
-                if let Some(d) = extract_doc(&attr) {
-                    if !doc.is_empty() {
-                        doc.push(' ');
-                    }
-                    doc.push_str(&d);
-                }
-            }
-
-            // Try to parse variant name
-            let name: Ident = match iter.parse() {
-                Ok(n) => n,
-                Err(_) => break, // No more variants
-            };
-
-            // Determine variant kind
-            let fields = if let Ok(group) = iter.parse::<ParenthesisGroupContaining<TokenStream2>>()
-            {
-                // Tuple variant - count fields by counting commas + 1
-                let content = group.content.to_string();
-                let field_count = if content.trim().is_empty() {
-                    0
-                } else {
-                    content.matches(',').count() + 1
-                };
-                VariantFields::Tuple(field_count)
-            } else if let Ok(group) = iter.parse::<BraceGroupContaining<TokenStream2>>() {
-                // Struct variant - extract field names
-                let mut field_names = Vec::new();
-                let mut field_iter = group.content.to_token_iter();
-                loop {
-                    // Skip attributes on fields
-                    while field_iter.parse::<Attribute>().is_ok() {}
-
-                    let field_name: Ident = match field_iter.parse() {
-                        Ok(n) => n,
-                        Err(_) => break,
-                    };
-                    field_names.push(field_name);
-
-                    // Skip : Type
-                    let _ = field_iter.parse::<Colon>();
-                    while field_iter.parse::<TokenTree>().is_ok() {
-                        // Consume type tokens until comma or end
-                        if field_iter.parse::<Comma>().is_ok() {
-                            break;
-                        }
-                    }
-                }
-                VariantFields::Named(field_names)
-            } else {
-                VariantFields::Unit
-            };
-
-            variants.push(ParsedVariant { name, doc, fields });
-
-            // Skip comma between variants
-            let _ = iter.parse::<Comma>();
-        }
-
-        Ok(variants)
-    }
-}
-
-/// Extract doc comment text from an attribute
-fn extract_doc(attr: &Attribute) -> Option<String> {
-    let content = attr.body.content.to_string();
-    // Check if it starts with "doc"
-    if content.starts_with("doc") {
-        // Extract the string after `doc = `
-        if let Some(idx) = content.find('=') {
-            let rest = content[idx + 1..].trim();
-            // Remove surrounding quotes
-            let unquoted = rest.trim_matches('"').trim();
-            return Some(unquoted.to_string());
-        }
-    }
-    None
-}
-
 /// The main derive macro - this is what users write `#[derive(FacetPoc)]` for.
 #[proc_macro_derive(FacetPoc, attributes(facet_poc))]
 pub fn derive_facet_poc(input: TokenStream) -> TokenStream {
@@ -315,7 +125,6 @@ pub fn derive_facet_poc(input: TokenStream) -> TokenStream {
     while let Ok(attr) = iter.parse::<Attribute>() {
         let content = attr.body.content.to_string();
         if content.starts_with("facet_poc") {
-            // Extract plugin names from facet_poc(plugin1, plugin2, ...)
             if let Some(start) = content.find('(') {
                 if let Some(end) = content.rfind(')') {
                     let inner = &content[start + 1..end];
@@ -342,7 +151,6 @@ pub fn derive_facet_poc(input: TokenStream) -> TokenStream {
     }
 
     if plugins.is_empty() {
-        // No plugins - go straight to finalize
         return quote! {
             ::facet_plugin_poc::__facet_finalize! {
                 @tokens { #input_tokens }
@@ -440,7 +248,7 @@ pub fn __plugin_debug_invoke(input: TokenStream) -> TokenStream {
     chain_next(invoke, "debug").into()
 }
 
-/// The finalizer - parses and generates code
+/// The finalizer - parses using facet-macro-parse and generates code
 #[proc_macro]
 pub fn __facet_finalize(input: TokenStream) -> TokenStream {
     let invoke = match PluginInvoke::parse(input.into()) {
@@ -448,12 +256,13 @@ pub fn __facet_finalize(input: TokenStream) -> TokenStream {
         Err(e) => return quote! { compile_error!(#e); }.into(),
     };
 
-    let parsed = match ParsedType::parse(invoke.tokens) {
+    // Parse the type using facet-macro-parse
+    let parsed = match parse_type(invoke.tokens) {
         Ok(p) => p,
         Err(e) => return quote! { compile_error!(#e); }.into(),
     };
 
-    let name = &parsed.name;
+    let name = parsed.name();
     let mut generated = TokenStream2::new();
 
     for plugin in &invoke.plugins {
@@ -479,116 +288,65 @@ pub fn __facet_finalize(input: TokenStream) -> TokenStream {
     generated.into()
 }
 
-/// Generate Display impl from doc comments using the template system
-fn generate_display(parsed: &ParsedType) -> TokenStream2 {
-    use std::collections::HashMap;
-    use template::{EvalContext, Value};
+/// Generate Display impl using the parsed type from facet-macro-parse
+fn generate_display(parsed: &PType) -> TokenStream2 {
+    let name = parsed.name();
 
-    let name = &parsed.name;
-
-    match &parsed.kind {
-        TypeKind::Enum(variants) => {
-            // Build the evaluation context
-            let mut ctx = EvalContext::new();
-            ctx.set("Self", Value::Tokens(quote! { #name }));
-
-            // Build variant list for the template
-            let variant_values: Vec<Value> = variants
+    match parsed {
+        PType::Enum(e) => {
+            // Build match arms for each variant
+            let match_arms: Vec<_> = e
+                .variants
                 .iter()
                 .map(|v| {
-                    let mut obj = HashMap::new();
-                    obj.insert("name".to_string(), Value::Tokens(quote! { #(v.name) }));
-                    obj.insert(
-                        "doc".to_string(),
-                        Value::String(if v.doc.is_empty() {
-                            v.name.to_string()
-                        } else {
-                            v.doc.clone()
-                        }),
-                    );
-
-                    // Add field info
-                    let (is_unit, is_tuple, is_named) = match &v.fields {
-                        VariantFields::Unit => (true, false, false),
-                        VariantFields::Tuple(_) => (false, true, false),
-                        VariantFields::Named(_) => (false, false, true),
-                    };
-                    obj.insert("is_unit".to_string(), Value::Bool(is_unit));
-                    obj.insert("is_tuple".to_string(), Value::Bool(is_tuple));
-                    obj.insert("is_named".to_string(), Value::Bool(is_named));
-
-                    // Add pattern for matching
-                    let pattern = match &v.fields {
-                        VariantFields::Unit => quote! {},
-                        VariantFields::Tuple(count) => {
-                            let field_names: Vec<_> = (0..*count)
-                                .map(|i| {
-                                    proc_macro2::Ident::new(
-                                        &format!("v{i}"),
-                                        proc_macro2::Span::call_site(),
-                                    )
-                                })
-                                .collect();
-                            quote! { ( #(#field_names),* ) }
-                        }
-                        VariantFields::Named(field_names) => {
-                            quote! { { #(#field_names),* } }
-                        }
-                    };
-                    obj.insert("pattern".to_string(), Value::Tokens(pattern));
-
-                    Value::Object(obj)
-                })
-                .collect();
-
-            ctx.set("variants", Value::List(variant_values));
-
-            // For now, we still use quote! for the final output since we can't easily
-            // represent the full match arm generation in the template yet.
-            // This is a stepping stone — the template handles the overall structure,
-            // but complex pattern matching still uses quote! directly.
-            //
-            // TODO: Extend template language to handle this more elegantly
-            let match_arms: Vec<_> = variants
-                .iter()
-                .map(|v| {
-                    let variant_name = &v.name;
-                    let format_str = if v.doc.is_empty() {
-                        variant_name.to_string()
+                    let variant_name = v.raw_ident();
+                    let doc = v.doc();
+                    let format_str = if doc.is_empty() {
+                        v.effective_name().to_string()
                     } else {
-                        v.doc.clone()
+                        doc
                     };
 
-                    let (pattern, format_args) = match &v.fields {
-                        VariantFields::Unit => (quote! {}, quote! {}),
-                        VariantFields::Tuple(count) => {
-                            let field_names: Vec<_> = (0..*count)
-                                .map(|i| {
-                                    proc_macro2::Ident::new(
-                                        &format!("v{i}"),
-                                        proc_macro2::Span::call_site(),
-                                    )
-                                })
-                                .collect();
-                            let pattern = quote! { ( #(#field_names),* ) };
-                            let has_positional = format_str.contains("{0}");
-                            let args = if has_positional {
-                                let args: Vec<_> =
-                                    field_names.iter().map(|n| quote! { , #n }).collect();
-                                quote! { #(#args)* }
-                            } else {
-                                quote! {}
-                            };
-                            (pattern, args)
+                    // Generate pattern based on variant kind
+                    if v.kind.is_unit() {
+                        quote! {
+                            Self::#variant_name => write!(f, #format_str)
                         }
-                        VariantFields::Named(field_names) => {
-                            let pattern = quote! { { #(#field_names),* } };
-                            (pattern, quote! {})
-                        }
-                    };
+                    } else if v.kind.is_tuple() {
+                        let field_count = v.kind.fields().len();
+                        let field_names: Vec<_> = (0..field_count)
+                            .map(|i| {
+                                proc_macro2::Ident::new(&format!("v{i}"), proc_macro2::Span::call_site())
+                            })
+                            .collect();
 
-                    quote! {
-                        Self::#variant_name #pattern => write!(f, #format_str #format_args)
+                        // Check if format string uses positional args
+                        let has_positional = format_str.contains("{0}");
+                        let format_args = if has_positional && !field_names.is_empty() {
+                            let args: Vec<_> = field_names.iter().map(|n| quote! { , #n }).collect();
+                            quote! { #(#args)* }
+                        } else {
+                            quote! {}
+                        };
+
+                        quote! {
+                            Self::#variant_name( #(#field_names),* ) => write!(f, #format_str #format_args)
+                        }
+                    } else {
+                        // Struct variant
+                        let field_names: Vec<_> = v
+                            .kind
+                            .fields()
+                            .iter()
+                            .map(|f| {
+                                let raw = f.raw_ident();
+                                quote! { #raw }
+                            })
+                            .collect();
+
+                        quote! {
+                            Self::#variant_name { #(#field_names),* } => write!(f, #format_str)
+                        }
                     }
                 })
                 .collect();
@@ -603,29 +361,12 @@ fn generate_display(parsed: &ParsedType) -> TokenStream2 {
                 }
             }
         }
-        TypeKind::Struct { doc } => {
-            // For structs, we demonstrate the template concept but fall back to quote!
-            // since the template system is still evolving.
-            //
-            // The context building shows what data would be available to templates:
-            let mut ctx = EvalContext::new();
-            ctx.set("Self", Value::Tokens(quote! { #name }));
-            ctx.set(
-                "doc",
-                Value::String(if doc.is_empty() {
-                    name.to_string()
-                } else {
-                    doc.clone()
-                }),
-            );
-            // In a full implementation, plugins would ship templates that use this context.
-            // For now, we generate code directly:
-            let _ = ctx; // silence unused warning
-
+        PType::Struct(s) => {
+            let doc = s.doc();
             let format_str = if doc.is_empty() {
                 name.to_string()
             } else {
-                doc.clone()
+                doc
             };
 
             quote! {
